slowly integrating ntorv3
