+++ crates/o5/src/common/pq_obfs/keys.rs
@@ -179,4 +179,73 @@ mod test {
             hex::encode(shared_secret_alice)
         );
     }
+
+    #[test]
+    fn mlkem1024_x25519_handshake_flow() {
+        let mut rng = rand::thread_rng();
+        // long-term server id and keys
+        let server_id_keys = HybridKey::new(&mut rng);
+        let server_id_pub = server_id_keys.get_public();
+        // let server_id = ID::new();
+
+        // client open session, generating the associated ephemeral keys
+        let client_session = HybridKey::new(&mut rng);
+
+        // client sends kyber25519 session pubkey(s)
+        let cpk = client_session.get_public();
+
+        // server computes kyberx25519 combined shared secret
+        let server_session = HybridKey::new(&mut rng);
+        let server_hs_res = server_handshake(&server_session, &cpk, &server_id_keys, &server_id);
+
+        // server sends kyberx25519 session pubkey(s)
+        let spk = client_session.get_public();
+
+        // client computes kyberx25519 combined shared secret
+        let client_hs_res = client_handshake(&client_session, &spk, &server_id_pub, &server_id);
+
+        assert_ne!(client_hs_res.is_some().unwrap_u8(), 0);
+        assert_ne!(server_hs_res.is_some().unwrap_u8(), 0);
+
+        let chsres = client_hs_res.unwrap();
+        let shsres = server_hs_res.unwrap();
+        assert_eq!(chsres.key_seed, shsres.key_seed);
+        assert_eq!(&chsres.auth, &shsres.auth);
+    }
+
+    #[test]
+    fn kyber_handshake_supplement_flow() {
+        let mut rng = rand::thread_rng();
+        // long-term server id and keys
+        let server_id_keys = HybridKey::new(&mut rng);
+        let server_id_pub = server_id_keys.get_public();
+        // let server_id = ID::new();
+
+        // client open session, generating the associated ephemeral keys
+        let client_session = HybridKey::new(&mut rng);
+
+        // client sends ed25519 session pubkey elligator2 encoded and includes
+        // session Kyber1024Supplement CryptoOffer.
+        let cpk = client_session.public_key();
+
+        // server computes KyberX25519 combined shared secret
+        let server_session = HybridKey::new(&mut rng);
+        let server_hs_res = server_handshake(&server_session, &cpk, &server_id_keys, &server_id);
+
+        // server sends ed25519 session pubkey elligator2 encoded and includes
+        // session Kyber1024Supplement CryptoAccept.
+        let spk = server_session.public_key();
+
+        // client computes KyberX25519 combined shared secret
+        let client_hs_res = client_handshake(&client_session, &spk, &server_id_pub, &server_id);
+
+        assert_ne!(client_hs_res.is_some().unwrap_u8(), 0);
+        assert_ne!(server_hs_res.is_some().unwrap_u8(), 0);
+
+        let chsres = client_hs_res.unwrap();
+        let shsres = server_hs_res.unwrap();
+        assert_eq!(chsres.key_seed, shsres.key_seed);
+        assert_eq!(&chsres.auth, &shsres.auth);
+    }
 }
+
